地址列表 删除地址完善

+++ src/pages/account/address/addressList/index.vue
@@ -3,24 +3,25 @@
     <li class="list" v-for="item in addressList" :key="item.ID">
       <div class="div">
         <div class="userInfo">
-          <p> {{item.ConsigneeName}} </p>
-          <p class='phone'>{{item.ContactMobile}}</p>
-          <p class="default" v-if="item.IsDefault"> 默认 </p>
+          <p>{{item.ConsigneeName}}</p>
+          <p class="phone">{{item.ContactMobile}}</p>
+          <p class="default" v-if="item.IsDefault">默认</p>
           <!-- <p class="tag" v-if="item.Tag != null">{{item.Tag}}</p> -->
         </div>
         <div class="address">{{item.Address + item.Address}}</div>
         <div class="editBtns">
           <div @click="doSetDefaultAddress(item.ID)">
-            <radio class="radio" :disabled="true" color="#cab894" :checked="item.IsDefault"/>设置默认
+            <radio class="radio" :disabled="true" color="#cab894" :checked="item.IsDefault" />设置默认
           </div>
           <div class="right">
-            <a class="editItem" :href="'/pages/account/address/editAddress/main?isEdit=true&consigneeId=' + item.ID" >
-              <Img class="imgIcon" src="https://pic.keede.com/app/images/icon_cart.png"></Img>
-              编辑 
+            <a
+              class="editItem"
+              :href="'/pages/account/address/editAddress/main?isEdit=true&consigneeId=' + item.ID"
+            >
+              <Img class="imgIcon" src="https://pic.keede.com/app/images/icon_cart.png"></Img>编辑
             </a>
-            <div class="editItem" @click="doDeleteAddress(item.ID)"> 
-              <Img class="imgIcon" src="https://pic.keede.com/app/images/icon_cart.png"></Img>
-              删除 
+            <div class="editItem" @click="doDeleteAddress(item.ID)">
+              <Img class="imgIcon" src="https://pic.keede.com/app/images/icon_cart.png"></Img>删除
             </div>
           </div>
         </div>
@@ -53,6 +54,7 @@ export default {
       api.getAddressList().then(({ Data }) => {
         //重新整理数据
         this.addressList = [...Data];
+        console.log('删除后刷新数据')
       });
       if (this.addressList.leng) {
         this.isNoData = true;
@@ -70,21 +72,33 @@ export default {
             icon: 'none',
           }); 
         }
-
       });
     },
     doDeleteAddress(consigneeId){
-      api.doDeleteAddress(consigneeId).then(({ Data, State, Msg }) => {
-        let tempMsg = "";
-        if (State){
-          this.getListData();
-        }else{
-          wx.showToast({
-            title: tempMsg,
-            icon: 'none',
-          }); 
+      let _this = this
+      wx.showModal({
+        title: '提示',
+        content: '确定删除收货地址？',
+        success (res) {
+          if (res.confirm) {
+            console.log('用户点击确定')
+            api.doDeleteAddress(consigneeId).then(({ Data, State, Msg }) => {
+              let tempMsg = "";
+              if (State){
+                console.log('删除---成功----')
+                _this.getListData();
+              }else{
+                wx.showToast({
+                title: tempMsg,
+                icon: 'none',
+                }); 
+              }
+            });
+          } else if (res.cancel) {
+            console.log('用户点击取消')
+          }
         }
-      });
+      })
     }
   }
 };
@@ -103,7 +117,7 @@ page {
   font-size: 28rpx;
   display: flex;
   align-items: center;
-  .div{
+  .div {
     width: 100%;
   }
 }
@@ -142,7 +156,7 @@ page {
   color: #717171;
   font-size: 24rpx;
   border-bottom: 1rpx solid #dcdcdc;
-  width: 100%
+  width: 100%;
 }
 .editBtns {
   height: 60rpx;
@@ -163,14 +177,13 @@ page {
   }
 }
 .radio {
-  transform:scale(0.7);
+  transform: scale(0.7);
   color: #cab894;
 }
 .imgIcon {
   width: 20rpx;
   height: 20rpx;
 }
-
 
 /*新建收货地址按钮*/
 .bottom-btn-box {
